requirements.txt
fastapi==0.110.3
google-generativeai==0.8.3
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn==0.30.6
streamlit==1.40.1
requests==2.32.3

